refactor: padronizar exports da presentation
