+++ src/presentation/contracts/controller.ts
@@ -1,4 +1,4 @@
-import { HttpResponse } from '@/presentation/contracts/http-response';
+import { HttpResponse } from '@/presentation/contracts';
 
 export interface Controller<T = any> {
   handle(request: T): Promise<HttpResponse>;
