+++ src/presentation/validation/index.ts
@@ -1,0 +1,2 @@
+export * from './required-fields-validator';
+export * from './validation-composite';
